refactor: Move start logic into own file
